Fix GPS location failing too fast
On Android 12 and above, we request location updates every 2 seconds.
These will be null if a location is not yet known.

Before this change, we aborted on the first such null value,
and fell back to "lastKnownLocation", which could also be null.

With this change, we skip over these null values and wait longer.