Fix crash server registration when multiple UP distributors are installed
The problem was that UP tries to show a dialog to let the user select which distributor to use.
However, this use the AddAccountActivity as a context.
But this activity was being finished, and the window was gone, causing a crash.

To fix this, don't register for UP in non-UI context (i.e. not the BootReceiver).
In the FMD Server registration case, we already redirect to the FMDServerActivity,
which itself also handles UP registration.