Replace PIN-hash-in-settings with PIN-in-enc-settings
Storing the cleartext PIN in encrypted settings improves performance,
since we no longer need the expensive hashing.

This eliminates the possiblity of a "DoS attack", where an attacker sends the
victim many messages with a dummy PIN, causing the victim to waste battery doing PIN hashing.

It does come at the cost that we can no longer backup the PIN, since the backup is currently unprotected.