Start services in Application::onCreate instead of MainActivity::onResume
This avoid excessive service starts, e.g. as reported in
https://gitlab.com/Nulide/findmydevice/-/issues/310.

onResume is called _a lot_, since currently all screens are activities.
So as the users navigates the app, they often change between activities.

Starting the services so often was originally intended as a reliability measure.
However, let's step back and err on the side of less excessiveness.

Starting the services once when the application starts _should_ be enough.