Fix: properly cleanup location when JobService is stopped by the system
It can happen that the system calls FmdJobService.onStopJob(),
to indicate that the job should stop running.
In that method, we (rightfully) call coroutineScope.cancel().

However, this causes the existing cleanup code in the LocateCommand to not run.

This has two effects:

1. locOnOffHandler.removeJob(res.jobId) is not called. Thus, the location is not turned off.
2. jobFinished() is not called. This is a problem for ServerLocationUploadService, see below.

Theoretically, ServerLocationUploadService.onStopJob() returns true, meaning that
the system should retry the job. However, sometimes it doesn't.
(This _might_ be related to the reuse of JOB_ID, not 100% sure.)
By calling jobFinished() in the cleanupHandler, we ensure that we instead
finish the job and call scheduleNextOccurrence().