ring: restore interruption filter
Since Android N, RINGER_MODE_SILENT and DND are closely coupled:
https://stackoverflow.com/questions/58044974/enable-silent-mode-in-android-without-triggering-do-not-disturb

Before this change:

- Before "silent & DND off" ==> After "silent & DND on"

Fixes https://gitlab.com/fmd-foss/fmd-android/-/issues/345
in the sense that with this change:

- Before "silent & DND off" ==> After "silent & DND off"

However, restoring the DND mode does **not** work:

- Before "silent & DND on" ==> After "silent & DND off"

This seems to be intentional by Android?
Only users (but not apps) are allowed to set the DND mode?