Refactor writeAsJson
The shared util.writeAsJson() receives a Gson object.
Let every repository provide a writeAsJson(), that uses the util function with its own Gson object.

This is necessary, because some Gson objects may be configured differently.

For example, the upcoming changes to the settings to store location accuracy/altitude/etc
as Doubles, and in particular as NaN, requires the Gson object of the SettingsRepository
to be configured differently to allow NaN.