Nudge users to enable the connectivity check
The connectivity check is off by default.
And many users don't enable it, because they don't know about it.
This came up as a topic on Matrix.

Due to privacy concerns, I want to keep it disabled by default.
(It allows a bad server to see all the IP addresses that your phone has as you go about your day.)

This commit adds showing a notification upon logging in/registering and upon app updates.
It also adds this topic as a "recommendation" to the "Setup Warnings" screen.
This change should 1) educate the user that this even exists, and 2) provide a non-intrusive nudge.

Alternatives considered:

Another idea was to enable the connectivity check by default on "self-hosted" servers.
The problem with this is that it is not a simple "is it fmd.nulide.de?".
For example, your local hacker space could host a public FMD Server.
So what we actually mean is not "self-hosted" but "private" servers.
And because it is hard to determine whether a server is public or private,
I prefer to keep the conn check disabled by default and let the user decide.