Add warning about criticality of choosing a proper password
I am intentionally saying "that you can recover" instead of remember.
Because you could choose a strong password that you cannot remember,
but you write it down and store it in a safe place
(e.g., the password manager on your laptop, or a paper that you put under your pillow, etc.).

This is in response to a discussion on Matrix, with a user asking
if the server administrator can help with a forgotton password.