Download command from FMD Server when app is opened or server settings resumed
Before this change, commands where only downloaded when a push was received.
Forcing command downloads upon app open is slightly useful for debugging
("at least the basic connection works").

Downside: this leaks to the server when the user opens the app (time, IP).
But: the server version check already leaks this info.
And people open FMD rarely, and when they do it is often to debug something...

Do it in:

- MainActivity.onCreate() => baseline, does not run when returning to app from recents.
  This limits noise.
- FMDServerActivity.onResume() => additionally, runs when returning from recents.
  Potentially more noise (in logs and more network requests).
  But if the user is in those settings, this is especially relevant.