Disable settings export in UpdateboardingModernCryptoActivity
This will clash after the export has been migrated to Kotlin coroutines.
And to keep things simple/small for now, I don't want to convert this to Kotlin,
but also not delete this update boarding (yet).

Thus as a minor inconvenience, disable the export.
This file is from September 2023 (two years ago), so nobody should be using this update boarding anyway.