Centralize job start/stop logging
Before this change, the FmdJobService already logged start/stop events together with the job ID.
However, the tag was a generic "FmdJobService".
So users had to know that JOB_ID=110 means "FmdBatteryLowService".

With this change, the TAGs are meaningful.
This allows us to remove the custom, more descriptive start/stop logs in some jobs.
This makes the logs less verbose and easier to skim.